+++ CHANGELOG.md
@@ -7,10 +7,13 @@
 - Global env variables
 - Add describe flag to run cmd to suppress command information
 - Add progress spinner for run/exec cmd
+- Add sub-commands
+- Run multiple commands
 
 ### Changed
 
 - [BREAKING CHANGE]: Renamed args in command block to env
+- [BREAKING CHANGE]: Renamed commands in root block to tasks
 - Env variables now support shell execution
 - Rename flag format to output when listing